Move sort type declarations to lib/tweet/types.ts
