+++ src/lib/tweet/sort-tweets.ts
@@ -1,18 +1,4 @@
-import { DeletedTweets, Tweet } from './types';
-
-export type TweetSortKey = 'created_time' | 'saved_time' | 'username';
-export type DeletedTweetsSortKey = 'deleted_time';
-export type SortOrder = 'asc' | 'desc';
-
-export interface TweetSort {
-  key: TweetSortKey;
-  order: SortOrder;
-}
-
-export interface DeletedTweetsSort {
-  key: DeletedTweetsSortKey;
-  order: SortOrder;
-}
+import { DeletedTweets, DeletedTweetsSort, Tweet, TweetSort } from './types';
 
 export const tweetSortFunction = ({
   key,
